Fix optimizedFCallBuiltin() name handling
Summary: optimizedFCallBuiltin() was using name() to retrieve the function's name, not fullName(). This means if you have a builtin static class function which has the same name as a top-level builtin, we'd optimize it as if it was the top-level builtin. For instance, IntlChar::chr() would be simplified as if it was chr(). Use fullName() instead. Note that optimizedFCallBuiltin() is now misnamed because we're optimizing more than FCallBuiltins (like in this case).

Reviewed By: swtaarrs

Differential Revision: D5021899

fbshipit-source-id: 157f5dd1ea70b06a517c77dedcef8f5c71d8562f