Catch a few more misuses of Awaitables
Summary:
The existing `enforce_not_awaitable` complains if the type is exactly `Awaitable<...>` or the unresolved type containing exactly one `Awaitable<...>` but not for an unresolved type containing multiple `Awaitable<...>`s (e.g. `Awaitable<int> | Awaitable<string>`) or an `Awaitable<...>` and another type. This diff expands the check to fire for any type we know can be an `Awaitable<...>` (even if it can be another type as well).

Additionally it introduces an `enforce_nullable_or_not_awaitable` helper that only runs `enforce_not_awaitable` if the type is not optional. This lets us catch a truthy check on an `Awaitable<bool> | bool` while allowing it for `Awaitable<bool> | ?_` etc.

Reviewed By: andrewjkennedy

Differential Revision: D4966279

fbshipit-source-id: 4f71cf5620eb080067d5dad3201694730aa87c95