Add HH\array_key_cast
Summary:
Add a new builtin called HH\array_key_cast. This builtin takes a value and
return what that value would be cast to if used as an array key. The exceptions
are arrays, vecs, dicts, keysets, and objects. The semantics of using these as
array keys cannot be described by a simple cast, so this builtin throws on those
inputs instead. Try to optimize away the builtin in both HHBBC and the JIT.

For testing, add a new intrinsic builtin called "launder_value". Its just the
identity function, but is deliberately not optimized in any way. It can be used
to prevent certain optimizations which might be undesirable in unit tests.

(Note: this ignores all push blocking failures!)

Reviewed By: mxw

Differential Revision: D5013649

fbshipit-source-id: b3d60c68dd4897a1e889044f1b697e8d5fe1a616