Support parallel suspension points within a statement's expressions
Summary:
Supports parallel suspension points by storing intermediate results from coroutine calls as hoisted variables in the closure.

In general, something of this pattern will be required if we need to "remember" one or more intermediate results while waiting for a suspended function to complete. In cases where coroutine results do not need to be remembered through another coroutine suspension, this approach will not be necessary. We can work on optimizing for that case, but I figured that this would work for now to support `suspend` in a variety of locations within a statement.

Reviewed By: ericlippert

Differential Revision: D5006406

fbshipit-source-id: 5fc813d046d86fbdef17aadc3fa31fdbdf77bc94