Update more full-fidelity-parser rewriters
Summary:
We now have an aggregating pre-order rewriter similar to the post-order rewriters I added in my previous diff.

I've also added rewriters that have a simplified contract; the function which determines if a node is to be rewritten in this version returns an optional node. "None" means "leave it the same", and "Some new_node" means to do a replacement.

Reviewed By: michaeltingley

Differential Revision: D4998949

fbshipit-source-id: a610dc2fbc2fc5fd0f0cede712c7360878b969a9