vixl: Backport 32-bit immediate handling.
Summary:
When using negative signed values as immediate for logical or
move operations (e.g. And, Eor, etc.) then the conversion to
uint64_t (argument type of the immedate defined by the vixl API)
might end up with the upper 32 bits set.

This is problematic when using W-registers (32 bit only).

The current arm64 code solves the problem by using the
MSKTOP() function, which seems to be derived from
upstream vixl. This patch puts the code to HHVM's vixl,
which allows to remove MSKTOP() in vasm-arm.cpp.

Signed-off-by: Christoph Muellner <christoph.muellner@theobroma-systems.com>
Closes https://github.com/facebook/hhvm/pull/7808

Differential Revision: D5028530

Pulled By: mxw

fbshipit-source-id: 56c0780d91077a312a04ca9c5514e27d18acf294