Type annotations on calls of generics dependent functions
Summary:
Besides the type annotation on object creation, i.e.
```
$v = new Vector<ThisThingRightHere>();
```
we also want to allow type annotation on the call sites of functions that have
generic type parameters, i.e.
```
$x = foo<Bar, Baz>($qux);
```

This is all well and good, but when this expression occurs in the position of a
function's argument, the parse is ambiguous, i.e.
```
$x = quux(foo<Bar, Baz>($qux));
```
Now, is this a call to a function with two boolean arguments, or is this a call
to `quux` with the //result of calling `foo`//?

In this diff, the choice is made to try and parse this as a type annotated
call. If it succeeds, that's held to be what it says. If it fails, assume
multiple expressions.

Reviewed By: andrewjkennedy

Differential Revision: D5028089

fbshipit-source-id: f69797d09c411af844203d227c8d36f5f1137895