Hack codegen: consistent namespaces
Summary:
This is a first cut at getting Hack and PHP namespaces right in code generation. We
* Generate the correct instructions with fallbacks (e.g. Cns vs CnsU)
* Track namespaces through top-level definitions by introducing a new SetNamespaceEnv constructor into the AST
* Emit namespaces correctly in type hints

All class, function, property and constant identifiers are now managed in `Hhbc_id` with a signature hiding the string representations.

There are some tricky aspects of top-level constants defined using `define('name', ...)`
that we don't yet get right. But putting this up now for review.

Reviewed By: hubyrod

Differential Revision: D5020144

fbshipit-source-id: dec1d3ccd0371517d034362813d623fba48a2d44