Delete incorrect assertion in load-elim
Summary:
This assertion doesn't hold because, in general, for two jit::Types T1
and T2, it's not necessarily the case that (T1 & T2) is <= T1 or T2, so
the intersecting of tracked.knownType with inst.dst().type() that we do
before this assertion won't necessarily "refine" it in the sense of
shrinking it.

Just delete the assertion; it doesn't look like we actually depend on it
for correctness.

Reviewed By: markw65

Differential Revision: D5043581

fbshipit-source-id: 1ecaa1b353d3efff833c0fa15e4dcea1df296062