Stop dropping empty jeprof files on oom
Summary:
The memory manager was set up with the intent that if jemalloc
profiling was active, we would dump a profile on oom. It tested that
via jemalloc's "opt.prof" config, which worked at the time this code
was written, because we didn't set opt.prof unless we were going to
enable profiling.

Since then, we've changed the default malloc_conf so that opt.prof is
always set.

This checks prof_active too, so that we don't dump useless files
everywhere.

Reviewed By: edwinsmith

Differential Revision: D5058444

fbshipit-source-id: a413bb4d2c6c7e5edb7d87cd5dd9aa5c7ea0e377