Enqueue unknown objects during initialization, simplify trace phase.
Summary:
Objects with an unknown type index can only come from req::malloc().
Resources, native objects, and WaitHandles all are allocated with
a type index from getIndexForMalloc<T>.

Avoid creating a list of unknown objects; when we see them during
the init phase, mark them and put them on the worklist; this results
in handling them in the same order as before, since the worlist is
processed in LIFO order; they'll sit at the bottom of the list until
the rest of the heap is traced. (Although the order no longer matters).

Reviewed By: ricklavoie

Differential Revision: D5059333

fbshipit-source-id: 8b6d2a1c966c5d7f0b4344bcf98e0fb39c1d91d9