Fix stack overflow for silly-deep parse trees
Summary: Running `hh_single_compile` on HHVM's `parser_massive_concat_exp.php` quick test caused a stack overflow. This is due to a very deep parse tree (the test is one giant concat expression, where the spine is as deep as the expression is long). Translating a `Minimal` syntax tree to a `Positioned` syntax tree can be done in a fully tail recursive fashion, but only if we generate the code.

Reviewed By: hubyrod

Differential Revision: D5019749

fbshipit-source-id: 9c636fa76759328efba47115e6fc01f0f00cacbc