rework lexing and unescaping of triplequoted strings
Summary:
triplequoted strings used to be unescaped in the lexer, but this was the wrong thing to do,
as revealed in the recently-added even_more_exciting_default_values.php test. This diff changes the
lexer to leave triplequoted things alone, and adds a separate unescaping operation prior to trying
to parse them as attributes.

Reviewed By: hubyrod

Differential Revision: D5061589

fbshipit-source-id: 0aeee30bb67a6224577d2d755cf6c0e719a1d94d