"hh_server's version doesn't match the client's". -- SO TELL ME THE VERSION\!
Summary:
Currently in case of a version mismatch we get messages like this:

  hh_server's version doesn't match the client's, so it will exit.
  Going to launch a new one.
  Server launched with the following command:
    '/data/users/ljw/fbsource/fbcode/buck-out/dbg/gen/hphp/hack/src/hh_server/hh_server' '-d' '/data/users/ljw/www-hg' '--waiting-client' '11'
  Spawned typechecker (child pid=2112370)
  Logs will go to /tmp/hh_server/zSdatazSuserszSljwzSwww-hg.monitor_log

I've found it frustrating when I encounter these. It always wastes a load of time digging through logs. With this diff, it will display more information:

  hh_server '...path_to_hh_server' was launched 15 hours ago.

I hope that with the extra path information, it'll save time when diagnosing errors.

Reviewed By: alexchow

Differential Revision: D4967303

fbshipit-source-id: a79f3d8f39f3c5949e1a01c75f57067359273a63