disallow missing semicolon in group use declarations
Summary:
This is for parity with HHVM. This statement always must be terminated with a semicolon. The use list declaration doesn't have this issue because it asks for a terminal character of `Tsc`, but group use wants a `Trcb` and neglects to consume a semicolon after.

Fixes https://github.com/facebook/hhvm/issues/7667.

Differential Revision: D5022898

fbshipit-source-id: e962232c7f28d4702035eba0acb9f28060fe922f