Front-end support for PHP's `global` keyword
Summary:
PHP allows for "globalizing" variables, i.e. from a local scope, inject a variable into the global scope. This is done with the `global` keyword:
```
function foo(): void {
  global $x;
}
```

The Hack tools did not have an internal representation for this, even though the Full Fidelity Parser already dealt with this correctly.

Reviewed By: oulgen

Differential Revision: D5060731

fbshipit-source-id: 966c55e8a792a5267ea229b5aa71072ddbc25640