Add fallback comparer for large lists in semdiff
Summary:
When semdiff can't prove equality, it currently falls back on a quadratic-time edit distance function. Some functions/methods
are jolly big, so that doesn't work so well. This diff adds a really dumb, linear-time pointwise comparer that is used when the
sizes get big. It also truncates the printing of detailed diff information in such cases, as it's probably not something we want to
actually look at.

Reviewed By: hubyrod

Differential Revision: D5069772

fbshipit-source-id: 8c65490e70e60144aa78b00d1e90efd9e5f4ea27