Store end of handler offset when available
Summary:
Introduce EHEnt::m_end pointing to the end of exception handler. This field is
used only for Catch EHEnts and is optional. This diff populates it from
hhas parser when .try {} .catch {} is used and from emitter.cpp, but it is not
yet populated by hhbbc and hhas parser parsing the legacy .try_catch {}
directive. The field is not yet used by anything except of debug output.

Reviewed By: paulbiss

Differential Revision: D5054260

fbshipit-source-id: 9e220723b52f18224f2f8b6afc9c6dbc9af9ab7f