Adjust priority of assignment in expressions
Summary: in PHP priority of the assignments in expressions is bumped up when left hand side of the assignment is valid lvalue (variable name, member selection, subscript expression or scoped expression). We recognise these cases and reset the precedence of the parser before parsing right hand side of the assignment to make sure that it is fully consumed. One note is that new precedence of the assignment is still less than i.e. precedence of the '$' operator to make sure that $$a = 'x' will be parsed as ($$a) = 'x' and not as $($x = 'x').

Differential Revision: D5079777

fbshipit-source-id: b37ed5373463d75bd69f8f825f529e2e19257a37