Simplify away useless CheckNonNull instructions
Summary:
Add a simplification rule to remove CheckNonNull instructions where the input is
always Nullptr, or the input isn't unioned with Nullptr. The only complication
to this is that LdContResumeAddr returned a TCA, but could actually be null at
runtime. Fix this by having it properly return TCA | Nullptr. From now on, the
contract is, if a pointer-ish type can be possibly null, it must be unioned with
Nullptr (this was largely the case already).

Reviewed By: swtaarrs

Differential Revision: D5063197

fbshipit-source-id: d64683ca632f43743b3d3ce58e3a60be5105fdfe