Propagate Func* values from SpillFrame to Call in load-elim
Summary:
Improve memory-effects and load-elim to propagate a known Func* from SpillFrame
to its usage in Call/CallArray. This involves adding new alias-classes to
PureSpillFrame and CallEffects which mark where exactly they will be
loading/storing the Func*. Load-elim can then see if it is aware of a known
Func* at that location, and if it is, it can update the Call/CallArray's
extra-data to mark that the callee is known.

As a further optimization, change LdARFuncPtr and LdARCtx to be pure-loads
(which they are) and give them more accurate memory effects. This means that we
can eliminate the whole CheckRefs sequence that we need for an unknown callee if
we determine a known callee.

Reviewed By: markw65

Differential Revision: D5070133

fbshipit-source-id: a71fb716f81581e3ff04421cf7668440fd13747b