Don't type-pun when emitting NativeImpl
Summary:
When lowering a (non-inlined) NativeImpl bytecode into IR, we have pointers to
all of the builtin's parameters. For primitive types, we load the parameter
using the pointer and pass it in directly. However, for some unknown reason, if
the parameter is a bool or ref, we'll load it back as an integer.

We seem to get away with this type-punning now, but slight improvements to
load-elim expose this as a problem. Load-elim will realize that the parameter is
actually a bool, and when we load it as an integer, we intersect the types and
get Bottom. Since there's not a clear reason why we're doing this, simply remove
it. If the parameter is a bool or ref, load it as such.

Reviewed By: paulbiss

Differential Revision: D5070603

fbshipit-source-id: c013b687e482fe049bac44ee51b5df5018713558