Run load-elim multiple times
Summary:
Load-elim is not guaranteed to reach a fixed-point in one iteration. The primary
reason is because the optimization phase (which runs at the end) can cause the
types of SSATmps to change. This can in turn change the types of values in
memory. So, that means the analysis phase can now derive better information. It
might be possible to have load-elim attempt to track how the types of SSATmps
will change as a result of optimizations in its analysis phase. However, an
easier solution is to just run load-elim in a loop until a fixed-point is
reached. To prevent a lot of looping by pathological units, or even infinite
loops due to bugs, bound the number of iterations we'll do. If we reach the max
iteration limit, report a perf warning.

While we're here, add some JIT stats for the load-elim pass. The number of
iterations along with the count of various optimizations performed.

Reviewed By: paulbiss, swtaarrs

Differential Revision: D5071919

fbshipit-source-id: 377775b3b8126117aa27edadc62f119fa3e8cb57