Strange conversions to Hack arrays should throw
Summary:
Right now if you attempt a "strange" conversion to a Hack array (null, bool,
int, string, double, resource, or non-iterable objects), we'll raise a notice
and return an empty Hack array. This is strange and can lead to bugs. Since the
whole reason for Hack arrays is to remove gotchas, throw in these cases instead.

Reviewed By: paulbiss

Differential Revision: D5089321

fbshipit-source-id: c3fdbdd8e1e9b680c13f882d4916a380b432bab7