jit: arm64: Fix regression in async testcases (repo auth)
A recent PR (#7833/61b3b0d) addresses the problem, that the arm64
lowering/emitter code need to truncate the operands of the vasm
instructions cmpb{} and cmpbi{}. The actual problem is, that the
given values are not Vreg8 values, but treated as Vreg8 values
by follow-up code. The fix in the named PR is to insert a movtqb{}
whenever a type value is lowered.

As reported by @swalk-cavium this breaks some async tests when
running in repo authoritative mode.

This patch addresses this issue by clearing bits 9-31 of the
type value, but keep bits 32-63 to keep the AUX value in the
register (clearing that bits leads to the regression in repo
authoritative mode).

The patch relies on the fact that the type value will later
be interpreted as 32-bit (W) register, i.e. ignoring the AUX
value anyways. Therefore a cmpb{} and cmpbi{} will not see
the AUX value.

Thanks @mxw for the idea for this PR.

Signed-off-by: Christoph Muellner <christoph.muellner@theobroma-systems.com>