Eliminate plain req::malloc, use malloc_unk explicitly or something better
Summary:
The default type for req::malloc(size,type) is Unknown; we don't
call it very often, and we want to make it hard to get by accident.
So, make the type_scan::Index argument required, and add req::malloc_unk(size)
for when the type is truly unknown. Rename existing callsites.

Reviewed By: ricklavoie

Differential Revision: D5062869

fbshipit-source-id: 44674457f7ed9e879edded17ef9515a7689d9762