Compute correct aawh size, type_scan flexible-array member.
Summary:
In conservative mode, we passed sizeof(c_AwaitAllWaitHandle)
to the scanner, which incorrectly only scanned the first part
of the object, but not the children allocated off the end.

To fix it:
1. pass the correct size (heapSize()) to the scanner, which fixes
conservative mode.
2. Convert the TYPE_SCAN_CUSTOM_FIELD to TYPE_SCAN_FLEXIBLE_ARRAY_FIELD,
so the generated scanner accepts the larger size and scans m_children
(similar to what the custom scanner did).

Also refactor scanHeader() slightly to avoid redudnant kind() tests.

Reviewed By: jano, ricklavoie

Differential Revision: D5098010

fbshipit-source-id: 7cbfd3afb9bad78ee7352a6c28f2ade0c91d51d1