Fix unstable serialization in fb_compact_serialize()
Summary:
fb_compact_serialize() is currently broken for short arrays that have numeric keys that aren't inserted in order. The current implementation will use a LIST_MAP for these arrays, which results in inputs like:

  array(2 => 'a', 1 => 'b', 3 => 'c')

being serialized using FB_CS_SKIP, which has the keys in numeric order. When deserialized, the result is:

  array(1 => 'b', 2 => 'a', 3 => 'c')

These two arrays are not equivalent. Fix is to only use LIST_MAP encoding for arrays whose input keys are in numeric order.

Reviewed By: markw65

Differential Revision: D4995313

fbshipit-source-id: 67c81b2bb3225bbf80b889f4002f0dd0441c2f15