Allow unreachable FPI regions in hhas
Summary:
The assembler bans calls from unreachable positions as it requires the stack
offset to be known to emit an FPI region. This was fine when all hhas was hand
rolled but makes writing a compiler that emits hhas challenging.

In emitter.cpp whenever we're about to emit an unreachable instruction we mark
the stack as empty and pretend it's reachable. This does basically the same thing
but only for FPI regions.

Reviewed By: swtaarrs

Differential Revision: D5096465

fbshipit-source-id: b10cf5940c4a1a1cb18cc8ddf7303e1443fa9659