Add --at-char option to hackfmt
Summary:
Add the --at-char option, which takes a 0-indexed character offset and formats the chunk group containing that character. The start and end offsets (inclusive and exclusive, respectively) of that chunk group in the original file are printed before the formatted text, so that IDEs can replace the original range with the formatted text.

Useful for as-you-type formatting, where typing certain characters will trigger formatting at that character's offset.

Reviewed By: eshrews, arxanas

Differential Revision: D4972208

fbshipit-source-id: e0a934bbfa920b5e0c5d33234857b94f1483643d