Fix incorrect indentation in --diff mode
Summary:
Previously, --diff mode never touched any line which was not modified in the diff. This behavior is usually reasonable, but falls apart when the modified range bisects what would otherwise be a single line of formatted text.

For example, this diff:

```
@@ -1,4 +1,4 @@
 if (true) {
   $foo =
-    'bar';
+    'baz';
 }
```

would normally be formatted like so:

  if (true) {
    $foo = 'baz';
  }

But since --diff mode is unwilling to touch the line containing `$foo =`, it results in this output:

  if (true) {
    $foo =
   'baz';
  }

where `'baz';` is not indented because it has no split before it (one space is printed before it because normally that space would be needed to separate it from the assignment operator).

We could mitigate this problem by ensuring splits are always broken on wherever modified range boundaries fall, but this would make --diff mode produce slightly different results than whole-file formatting, which is undesirable.

Instead, we expand the diff's modified ranges to the nearest split which the line splitter actually decided to break on. This means that lines which were not modified in the diff will be affected, but we only touch the bare minimum required to ensure that --diff mode produces the same output as whole-file formatting.

Reviewed By: eshrews, arxanas

Differential Revision: D4974605

fbshipit-source-id: bb2bb3c0c6b63fa58669a64e4b6252fa0dfb5f58