Expand --at-char range to containing node
Summary:
Currently, --at-char expands the range to the chunk group containing the given position. This is an okay approximation, but we really want a more semantically useful unit. This changeset expands the beginning of the range to the start offset of the AST node which is the direct parent of the token at the given position, then expands the beginning of the range to the nearest line boundary in the original source (since we want to be able to ensure a leading newline before the formatted node if one should be there), then expands the range to the nearest split boundaries (since Line_splitter requires it). It does not expand the end of the range, since we don't want to modify any characters after the cursor in as-you-type formatting.

This changeset also ensures that we don't include a trailing newline in --at-char output. Previously, we did so because the Line_splitter requires ranges to begin and end at line boundaries, but this is undesirable for as-you-type formatting because the cursor is at the end of the node being formatted (so including a newline in the output places the cursor at the beginning of the next line).

Depends on D4974605

Reviewed By: arxanas

Differential Revision: D5079312

fbshipit-source-id: 35066fc291f9007688c7a78a82d52e35ac406429