Implement StringData::hashHelper in AArch64 assembly
Summary:
StringData::hashHelper() is implmented in x64 assembly and used
when NO_M_DATA is set, which in turn is controlled by USE_LOWPTR.
This patch provides an equivalent version for Aarch64.

There are no new errors in the unit tests, which were tested with
{-m interp, -m pgo, -r, and empty} and OSS Performance suite still
runs fine, both with and without USE_LOWPTR and ENABLE_AARCH64_CRC
defined.
Closes https://github.com/facebook/hhvm/pull/7809

Differential Revision: D5013451

Pulled By: mxw

fbshipit-source-id: ab6e3f5ba7ced9a909f6935a0eb63265a1b38c41