Delete type-conversions.h
Summary:
These functions are archaic and mostly unused---in most cases, the only
overloads used were the identity overload, the `const StringData*`
overload, and the `const Variant&` overload, and never in any generic
contexts.

Replace the whole thing with `double_to_int64(double)`, which takes the
place of the old `toInt64(double)`.

Reviewed By: jano

Differential Revision: D5116313

fbshipit-source-id: 5318c103dc136e0bc9ccbfe72f3fc63810ac8bf2