*easy* Add a flag "do_post_handoff_handshake" to clientConnect
Summary:
I want to re-use clientConnect for establishing persistent connection. But persistent connection needs to do the post-handoff handshake later on, asynchronously -- it can't be done as part of connecting. (The post-handoff handshake involves waiting for a "Hello" from the server, sending it the desired connection type, and waiting for its response).

This diff adds a flag to configure whether clientConnect.connect should do that post-handoff handshake or not.

Reviewed By: alexchow

Differential Revision: D5111117

fbshipit-source-id: f72312b85ab60f00391205b41e82b5ddcad7586a