Switch server "hello" from a string to a ServerCommandTypes
Summary:
Previously, we waited for a "hello" from the server before running the persistent connection and receiving responses+pushes from it.

This diff moves the "hello" into responses+pushes+hello. The reason is so that the persistent connection can have a single uniform "wait-for-server" message function to receive any of those things.

Why? -- it's because after we've done the handoff negotiation, the LSP server will be in a state where it can receive either a client message or a server hello message. Once it's received the hello, then it will be in a state where it can receive either a client message or a server response+push.

It makes the code easier if we can treat those two states uniformally. Basically, we no longer get type system support for enforcing that the hello comes first before any response+pushes, and we instead look at a "state" field to decide how to handle whichever server message arrives.

Reviewed By: alexchow

Differential Revision: D5111124

fbshipit-source-id: cd0ab94fe921f8cbeaa22ed1d5ce9bfa3a4b9e85