Add progress-reporting while we're waiting for hh_server
Summary:
When you do "hh_client" at the command-line, it shows a spinner to show which phase it's in of getting hh_server up and running. This diff adds the same functionality for when Nuclide has to start up hh_server.

At the moment, this diff provisionally writes status to the console. Kasper has produced a nicer status-bar indicator. Once he lands that work, then we'll change to us it.

Overall, the UX philosophy behind this change is "reassure the user that something useful is going on; don't give them just an unresponsive language service with no understanding about what's happening".

Note that we had to rewrite the get_messages code so it returns a Tick event after 1s timeout. This meant that the old IDLE code had to be rewritten too.

Reviewed By: arxanas

Differential Revision: D5111136

fbshipit-source-id: 9cfd9a75109413dec9a8648b5dd7a39a14a5c860