don't write out /etc/hh.conf to stderr
Summary:
When Nuclide calls into hh_client and does the "initialize" request, it keeps note of what was written to stderr. That's because this stuff is often useful to diagnose issues (particularly if hh_client failed to start).

The "on-type-formatting" incidentally caused a dump of /etc/hh.conf to be written to stderr even on normal startup. I don't think this is user-relevant information, so I want to suppress it.

Reviewed By: arxanas

Differential Revision: D5130826

fbshipit-source-id: a7d5385c67fac2176a8df34c8a08040cf8cba7dc