Decrease the cost of the split following an assignment operator
Summary:
Currently, the split after an assignment operator is more costly to break on than most other splits. I assume this was done to force hackfmt to produce this solution:

```
$value = function_call(
  $first,
  $second,
  $third,
  $fourth,
  $fifth,
  $sixth,
  $seventh,
);
```

When it would otherwise produce this solution:

```
$value =
  function_call($first, $second, $third, $fourth, $fifth, $sixth, $seventh);
```

However, I don't see why the first solution is more desirable than the second. Since we have had a user object to hackfmt preferring the first solution, this changeset simplifies hackfmt by getting rid of Cost.Assignment and allowing the second solution.

Reviewed By: arxanas

Differential Revision: D5113143

fbshipit-source-id: 04e6aa80f2c114ba1593ebb187e3032779517d96