Fix AsyncGeneratorWaitHandle refcount bug
Summary:
Unfinished AsyncGeneratorWaitHandles are always referenced by AsyncGenerator's
m_waitHandle property. This reference was conflated with the semi-implicit
"I am still running" reference implied by various states:

- `BLOCKED` - referenced by child
- `READY` - referenced by each entry in the AsioContext's runnable queue
- `RUNNING` - referenced by being executed

When transitioning between states, incref/decref pairs are simply avoided. The
initial incref happens at the construction and initial decref at transition to
the `SUCCEEDED` or `FAILED` state. This exactly matched the lifetime of the
AsyncGenerator::m_waitHandle reference, so we just relied on this reference to
maintain the object alive.

However, there is one edge case. If the AsyncGeneratorWaitHandle becomes in the
`READY` state outside of the asio context, it gets decrefd, removing the
implicit reference while the object is still being referenced by m_waitHandle.
This triggers destruction of AsyncGeneratorWaitHandle, dropping its reference
to the AsyncGenerator, which will eventually get destructed and try to drop the
reference to the already destructed AsyncGeneratorWaitHandle, causing a
segfault.

This situation may happen if an async execution subtree containing an async
generator gets abandoned, but the direct dependency of such async generator
gets finished (e.g. because of a <<__Memoize>>), followed by immediately
exiting the context.

This diff fixes that by adding a separate incref/decref pairs for the
semi-implicit refcount. When the above unlikely scenario happens, the pair of
AsyncGenerator and AsyncGeneratorWaitHandle will simply leak, as they will
continue referencing each other. We will rely on GC to eventually release
the memory.

Reviewed By: edwinsmith

Differential Revision: D5136085

fbshipit-source-id: d78729923ac001016f7efc852bc0b0543f2f9270