Rename deltas, comments, trace messages in refreshStatsImpl()
Summary:
In this function, the word "delta" is used two ways:
One set of "Delta" variables are the difference between
*m_allocated and *m_deallocated, which is the currently
allocated memory (aka usage). Rename to "usage".

The other is the difference between previous and current
counters, measured at two points in time. In this case,
just inline the subtraction, it's clearer this way.

Also clarify some TRACE messages and comments.

Reviewed By: alexeyt

Differential Revision: D5137929

fbshipit-source-id: 301a35e3a363aad606f172c530791ab069cf1222