Kill MemoryUsageStats::mallocDebt
Summary:
MemoryUsageStats::capacity tracks the capacity of the request
heap as the sum of slabs and large objects. We also have logic
to exclude this capacity from auxUsage, by accumulating a
mallocDebt counter as we grow & shrink the heap, and then
incorporating that counter into auxUsage incrementally.

Simplify this by eliminating the mallocDebt counter completely,
so that auxUsage includes capacity, and add a new accessor
to subtract capacity on the fly, as needed.

Rename auxUsage to threadUsage since it no longer excludes the
heap, and to make sure we use the new accessor.

Reviewed By: alexeyt

Differential Revision: D5139382

fbshipit-source-id: b1210210970190c18ad953b1606eecff6c9d224d