Hack codegen: add hphp/tests/slow to comparison tests
Summary:
Now that Hack codegen is getting more complete, let's add the hphp/test/slow tests to the files that we diff generated code.
This is important to improve coverage and prevent regressions; for example, there are no memoize tests in test/quick.

Added a separate blacklist file for test/slow, and renamed the blacklist file for the test/fast.

Reviewed By: arxanas, oulgen

Differential Revision: D5137368

fbshipit-source-id: 3b08fba2ba577e309de8a1dc7bf7e778100bef55