added a trivial case of associativity of concatenation
Summary:
Hack and HHVM sometimes generate code that concatenates strings in a different order. This adds
the equivalence String s; Concat; Concat = Concat; String s; Concat to semdiff. Note that this is not very
general - it only applies when the last thing to be concatenated is a string literal and nothing else
gets interleaved in the sequence of three instructions. But it does show up in examples.

Reviewed By: hubyrod

Differential Revision: D5144613

fbshipit-source-id: 9a36702c24239a3b0694306ab45ac2fb892ca792