Revamped schema format for `Syntax` generation
Summary:
In order to generate typed versions of parse trees, we need to inform the generator about the types to generate. For this, the schema used (list-of-list-of-string previously), we need a little more expressive power in the schema. This diff introduces `aggregate_types` and `child_spec` which allow us to express these type relations between different nodes in the parse tree. To keep the schema readable (admittedly subjective), this diff also inlines the conversion from the inner list-of-string things to records into the schema itself. This involves a lot of regex based transformation, so FYI, this is the general approach:

 * Replacing `  \\\\[ ("\\\\w+");\\\\n    ("\\\\w+");\\\\n  ("\\\\w+");\\\\n    ("\\\\w+");((\\\\n    "\\\\w+";)*\\\\n    "\\\\w+") \\\\];\\\\n` with `  ; { kind_name   = $1\\\\n    ; type_name   = $2\\\\n    ; func_name   = $2\\\\n    ; description = $3\\\\n    ; prefix      = $4\\\\n    ; aggregates  = []\\\\n    ; fields      = |$5|\\\\n    }\\\\n`
 * Replacing repeatedly `(= \\\\|[^\\\\|\\\\n]*)\\\\n` with `$1`
 * Replacing repeatedly `= \\\\|(( "\\\\w+";)*)\\\\s\\\\s+"` with `= |$1 "`
 * Replacing repeatedly `(= \\\\| "\\\\w+"(; "\\\\w+")*)(; "\\\\w+")` with `$1\\\\n      $3, Token`
 * Replacing `(      )(; "\\\\w+")\\\\|` with `$1$2\\\\n$1]`
 * Replacing `= \\\\| ("\\\\w+")\\\\|` with `= [ $1, Token ]`
 * Replacing `\\\\s+= \\\\| ("\\\\w+")\\\\n` with ` =\\\\n      [ $1, Token\\\\n`
 * Fill in proper spec (instead of `Token`) for the different `fields`.

Reviewed By: michaeltingley

Differential Revision: D5121325

fbshipit-source-id: 8e3997d44a7c629c233eb1c4efc32d8789cccb02